% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{capa_detection}
\alias{capa_detection}
\title{Function to detect point anomalies in a dataset using capa of the anomaly package}
\usage{
capa_detection(
  .data,
  .type = "robustmean",
  .min_seg_len = 2,
  .country_col = NULL,
  .time_col = NULL,
  .indicator_col = NULL
)
}
\arguments{
\item{.data}{A data frame containing the dataset to be analyzed.}

\item{.type}{A character string specifying the type of anomaly detection to be used, based on the
capa method in the anomaly package. Default is "meanvar", which is for collective anomalies using joint
changes in mean and variance.}

\item{.min_seg_len}{A numeric value specifying the minimum segment length for the capa method. Default is 2.}

\item{.country_col}{A character vector specifying the column names for country identifiers.}

\item{.time_col}{A character string specifying the column name for time identifiers.}

\item{.indicator_col}{A character string specifying the column name for indicator identifiers.}
}
\value{
A data frame containing the original dataset with an additional column indicating the point anomaly status.
}
\description{
This function detects point anomalies in a dataset using the capa method from the anomaly package.
}
