% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{convert_to_tsibble}
\alias{convert_to_tsibble}
\title{Convert data to tsibble}
\usage{
convert_to_tsibble(
  .data,
  .country_col,
  .time_col,
  .indicator_col,
  .frequency = "yearly",
  .long_format = TRUE
)
}
\arguments{
\item{.data}{A data frame containing the data to be converted.}

\item{.country_col}{A character vector specifying the name(s) of the country
identifiers.}

\item{.time_col}{A character vector specifying the name(s) of the time
identifiers. See \code{tsibble::as_tsibble()} for more details.}

\item{.indicator_col}{A character vector specifying the name(s) of the indicator
identifiers.}

\item{.frequency}{A character vector specifying the frequency of the time series.
Default is "yearly" for yearly data. The other values are "month" and "quarter".}

\item{.long_format}{A logical value indicating whether to return the data in long format.
Default is TRUE.}
}
\value{
A tsibble object with the specified country and time identifiers.
}
\description{
This function converts a data frame to a tsibble object.
}
