% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect}
\alias{detect}
\title{Function to apply outlier detection}
\usage{
detect(
  .x,
  .method = "zscore",
  .country_col = NULL,
  .time_col = NULL,
  .indicator_col = NULL,
  .additional_cols = FALSE,
  ...
)
}
\arguments{
\item{.x}{An object from the \code{normalize} function, which is a data frame containing the normalized
data to be analyzed.}

\item{.method}{A single character string or character vector specifying the outlier detection method to be used.
Options are \code{"zscore"}, \code{"tsoutlier"}, \code{"isotree"}, \code{"outliertree"}, or \code{"capa"}.
\code{"zscore"} is the default method, it calculates the Z-score for each observation and identifies outliers
based on a specified threshold (defaults to 3). \code{"tsoutlier"} uses the IQR method to detect outliers in
time series data, identifying values that are greater than 3 (default) times the IQR or less than -3 (default)
times the IQR (one can change this threhold). \code{"isotree"} applies the \code{isotree} algorithm of the
\code{isotree} package to detect outliers. Similarly, \code{"outliertree"} applies the \code{outliertree} algorithm
of the \code{outliertree} package to detect outliers. Finally, \code{"capa"} applies the \code{capa} method
of the \code{anomaly} package to detect point and collective anomalies.}

\item{.country_col}{A character vector specifying the column names for country identifiers. Not needed if these
are the same as the ones defined for the \code{normalize} function.}

\item{.time_col}{A character string specifying the column name for time identifiers.Not needed if these are the
same as the ones defined for the \code{normalize} function.}

\item{.indicator_col}{A character string specifying the column name for indicator identifiers. Not needed if these
are the same as the ones defined for the \code{normalize} function.}

\item{...}{Additional arguments to be passed to the specific outlier detection method. See Details
for arguments for each method.}
}
\value{
A data frame containing the original dataset with additional columns indicating the outlier status.
}
\description{
This function applies outlier detection to a given dataset using the specified method.
}
\details{
The function supports multiple outlier detection methods:
\itemize{
\item \code{"zscore"}: Detects outliers based on Z-scores. The threshold can be adjusted using the \code{.threshold} argument.
\item \code{"tsoutlier"}: Detects outliers using the IQR method. The threshold can be adjusted using the \code{.threshold} argument.
\item \code{"isotree"}: Applies the isotree algorithm. The number of dimensions, trees, threads, sample size, and threshold can be adjusted using
the \code{.ndim}, \code{.ntrees}, \code{.nthreads}, \code{.sample_size}, and \code{.threshold} arguments, respectively.
\item \code{"outliertree"}: Applies the outliertree algorithm. The columns to be used for outlier detection can be specified using
the \code{.cols} argument. The threshold can be adjusted using the \code{.threshold} argument.
\item \code{"capa"}: Applies the capa method from the anomaly package. The type of anomaly detection can be specified using the
\code{.type} argument, which can be "robustmean" (default), "mean", or "meanvar". The minimum segment length can be adjusted using the
\code{.min_seg_len} argument.
}
}
