% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{detrend}
\alias{detrend}
\title{Detrend a time series using a smoothing spline}
\usage{
detrend(.data, .year)
}
\arguments{
\item{.data}{A numeric vector representing the time series data to be detrended.}

\item{.time}{A numeric vector representing the time points corresponding to the data.}
}
\value{
A numeric vector of the same length as .data, containing the detrended values.
}
\description{
This function detrends a time series using a smoothing spline.
}
