% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values in a time series}
\usage{
impute_missing(
  .data,
  .value_col,
  .method = "na_interpolation",
  .country_col,
  .time_col,
  .indicator_col,
  ...
)
}
\arguments{
\item{.data}{A tsibble object containing the data to be imputed.}

\item{.value_col}{A character vector specifying the name of the value column.}

\item{.method}{A character vector specifying the imputation method to be used.
The default is \code{na_interpolation}. Other options are: na_kalman, na_mean,
na_locf, na_ma, na_seadec, na_random.}

\item{.country_col}{A character vector specifying the name(s) of the country
identifiers.}

\item{.time_col}{A character vector specifying the name(s) of the time
identifiers.}

\item{.indicator_col}{A character vector specifying the name(s) of the indicator}

\item{...}{Additional arguments to be passed to the imputation function.}
}
\value{
A tsibble object with the imputed data and a column identifying the imputed values.
}
\description{
This function imputes missing values in a time series using a function
from imputeTS package. The current options are: na_interpolation,
na_kalman, na_mean, na_locf, na_ma, na_seadec, na_random.
}
