% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{isotree_detection}
\alias{isotree_detection}
\title{Function to apply outlier detection}
\usage{
isotree_detection(
  .data,
  .cols = NULL,
  .ndim = 1,
  .ntrees = 10,
  .nthreads = 2,
  .threshold = 0.5,
  ...
)
}
\arguments{
\item{.data}{A data frame containing the dataset to be analyzed.}

\item{.ndim}{A numeric value specifying the number of dimensions for the isotree algorithm. Default is 1.}

\item{.ntrees}{A numeric value specifying the number of trees to be used in the isotree algorithm. Default is 10.}

\item{.nthreads}{A numeric value specifying the number of threads to be used in the isotree algorithm. Default is 2.}

\item{.threshold}{A numeric value specifying the threshold for outlier detection. Default is 0.5.}

\item{...}{Additional arguments to be passed to the isotree algorithm.}

\item{.method}{A string specifying the outlier detection method to be used. Options are "zscore", "iqr", or "mad".

Function to apply the isotree algorithm}
}
\value{
A data frame containing the original dataset with an additional column indicating the outlier score and indicator
based on the specified threshold.
}
\description{
This function applies outlier detection to a given dataset using the specified method.

This function applies the isotree algorithm to a given dataset for outlier detection.
}
