% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize and detrend the series for each country}
\usage{
normalize(
  .data,
  .value_col,
  .country_col,
  .time_col,
  .indicator_col,
  .frequency = "yearly",
  .detrend = TRUE,
  .season = NULL,
  .impute = TRUE,
  .method = "na_interpolation",
  .long_format = TRUE,
  .keep_decomp = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data frame containing the data to be normalized and detrended.}

\item{.value_col}{A character with the name of the column to be normalized and detrended,
can be integer refering to indexes of the columns (e.g., 8), or the name
of the column (e.g., "AG.CON.FERT.PT.ZS" or "value").}

\item{.country_col}{A character vector specifying the name(s) of the country
identifiers}

\item{.time_col}{A character vector specifying the name(s) of the time
identifiers.}

\item{.indicator_col}{A character vector specifying the name of the indicator
identifiers.}

\item{.frequency}{A character vector specifying the frequency of the time series.
Default is "yearly". The other values are "monthly" and "quarterly".}

\item{.detrend}{A logical value indicating whether to detrend the data. The
\code{STL} function from the \code{feasts} is used for smoothing. Default is TRUE.}

\item{.season}{A numeric value specifying the seasonality to be used. Default is NULL.
If NULL, the function will use the default seasonality of the data.}

\item{.impute}{A logical value indicating whether to impute missing values using the
specified method from \code{imputeTS}. Default is TRUE.}

\item{.method}{A character vector specifying the imputation method to be used.
The default is \code{na_interpolation}. Other options are: na_kalman, na_mean,
na_locf, na_ma, na_seadec, na_random. Check \code{imputeTS} for more details.}

\item{.long_format}{A logical value indicating whether to return the data in long format.
Default is TRUE.}

\item{.keep_decomp}{A logical value indicating whether to keep the decomposition
components in the output. Default is FALSE.}

\item{...}{Additional arguments to be passed to the imputation function.
See \code{imputeTS} for more details on the different methods.}
}
\value{
A data frame with the normalized and detrended data.
}
\description{
This function normalizes and detrends the series for each country.
}
