% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{outliertree_detection}
\alias{outliertree_detection}
\title{Function to detect outliers using the outliertree package}
\usage{
outliertree_detection(
  .data,
  .cols = NULL,
  .threshold = 0.5,
  .save_outliers = TRUE,
  .nthreads = 2,
  ...
)
}
\arguments{
\item{.data}{A data frame containing the dataset to be analyzed.}

\item{.cols}{A character vector specifying the columns to be used for outlier detection. Default is NULL, which uses all columns.}

\item{.threshold}{A numeric value specifying the threshold for outlier detection. Default is 0.5.}

\item{.save_outliers}{A logical value indicating whether to save the outliers. Default is TRUE.}

\item{.nthreads}{A numeric value specifying the number of threads to be used. Default is 2.}

\item{...}{Additional arguments to be passed to the outliertree function.}
}
\value{
A data frame containing the original dataset with an additional column indicating the outlier status.
}
\description{
This function detects outliers in a dataset using the outliertree package.
}
