% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{plot.maly_detect}
\alias{plot.maly_detect}
\title{Plot method for maly_detect class}
\usage{
\method{plot}{maly_detect}(x, country = NULL, indicator = NULL, x.lab = NULL, y.lab = NULL, ...)
}
\arguments{
\item{x}{An object of class maly_detect.}

\item{country}{A character vector specifying the country to plot.
Default is NULL, which means first country in the data will be used.}

\item{indicator}{A string specifying the indicator to plot. Only one per graph.
Default is NULL, which means the first indicator in the data will be used.}

\item{x.lab}{A string specifying the label for the x-axis.
Default is NULL, which means the time column will be used.}

\item{y.lab}{A string specifying the label for the y-axis of the original value (top graph).
Default is NULL, which means the value column will be used.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A ggplot object showing the outliers detected in the specified country and indicator.
}
\description{
This function provides a plot of the outliers detected in the maly_detect class object.
}
