% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{point_anomalies}
\alias{point_anomalies}
\title{Function to return point anomalies}
\usage{
point_anomalies(object, epoch = NULL)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{capa}}.}

\item{epoch}{Positive integer. CAPA methods are sequential and as such, can generate results up to, and including, any epoch within the data series. This can be controlled by the value
of \code{epoch} and is useful for examining how the inferred anomalies are modified as the data series grows. The default value for \code{epoch} is the length of the data series.}
}
\value{
A data frame.
}
\description{
Function to return point anomalies
}
