% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{summary.maly_detect}
\alias{summary.maly_detect}
\title{Summary method for maly_detect class}
\usage{
\method{summary}{maly_detect}(x, ...)
}
\arguments{
\item{x}{An object of class maly_detect.}

\item{...}{Additional arguments (not used).}
}
\description{
This function provides a summary of the maly_detect class object.
}
