% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{summary.maly_norm}
\alias{summary.maly_norm}
\title{Summary method for maly_norm class}
\usage{
\method{summary}{maly_norm}(object, ...)
}
\arguments{
\item{object}{A maly_norm object.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints the summary of the maly_norm object.
}
