% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wdi_download}
\alias{wdi_download}
\title{Function to download and prepare CSV file from the World Bank}
\usage{
wdi_download(.path = tempdir(), .prepare = TRUE, .additional = FALSE)
}
\arguments{
\item{.path}{The path where the CSV file will be downloaded. Default is the temporary directory.}

\item{.prepare}{A logical value indicating whether to prepare the data in wide format
with the following columns:
\itemize{
\item Country.Name
\item Country.Code
\item Year
\item Indicator Code as columns ...
}}

\item{.additional}{A logical value indicating whether to download additional files.
The additional files contain the information about the countries and series. Default is FALSE.
\itemize{
\item Series information
\item Series by year information
\item Country information
\item Country by series information
}}
}
\value{
Returns either a data frame in a tidy format (\code{.prepare = TRUE}),
a data frame with the data directory (\code{.prepare = FALSE}), or
a list with the data (either tidy or not) and additional data.frame with the
information provided in \code{.additional = TRUE}.
}
\description{
Function to download and prepare CSV file from the World Bank
}
