% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{write_to_csv}
\alias{write_to_csv}
\title{Function to save the maly_detect object to a file}
\usage{
write_to_csv(x, file, top_outliers = NULL, additional_cols = FALSE, ...)
}
\arguments{
\item{x}{An object of class maly_detect.}

\item{file}{A character string specifying the file path where the object should be saved.}

\item{top_outliers}{A integer value indicating whether to save only the top outliers.
#' Default is NULL, which saves all data. If only one method specified, it will use the Zscore
to order the data based on the absolute value of the Zscore. If multiple methods were used,
it will use the \code{outlier_indicator_total} column to order the data.}

\item{additional_cols}{A logical value indicating whether to save additional columns in the data frame.
These columns include the original columns, the additional columns from normalize and
the outlier indicator columns.}

\item{...}{Additional arguments to be passed to the write.csv function.}
}
\description{
This function saves the maly_detect object to a file in csv format.
}
