% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{zscore_detection}
\alias{zscore_detection}
\title{Function to detect outliers using the Zscore and ranking it}
\usage{
zscore_detection(.data, .threshold = 3)
}
\arguments{
\item{.data}{A data frame containing the dataset to be analyzed.}

\item{.threshold}{A numeric value specifying the Z-score threshold for outlier detection. Default is 3.}
}
\value{
A data frame containing the original dataset with an additional column indicating the outlier status
}
\description{
This function detects outliers in a dataset using the Z-score method and ranks them based on their Z-scores.
}
