% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{capa_detection}
\alias{capa_detection}
\title{Function to detect point anomalies in a dataset using capa of the anomaly package}
\usage{
capa_detection(
  .data,
  .type = "robustmean",
  .min_seg_len = 2,
  .country_col = c("Country.Code", "Country.Name"),
  .time_col = "Year",
  .indicator_col = "Indicator.Code"
)
}
\arguments{
\item{.data}{A data frame containing the dataset to be analyzed.}

\item{.type}{A character string specifying the type of anomaly detection to be used, based on the
capa method in the anomaly package. Default is "meanvar", which is for collective anomalies using joint
changes in mean and variance.}
}
\value{
A data frame containing the original dataset with an additional column indicating the point anomaly status.
}
\description{
This function detects point anomalies in a dataset using the capa method from the anomaly package.
}
