% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_tsibble}
\alias{convert_to_tsibble}
\title{Convert data to tsibble}
\usage{
convert_to_tsibble(
  .data,
  .country_col = c("Country.Code", "Country.Name"),
  .year_col = "Year",
  .indicator_col = "Indicator.Code",
  .frequency = "yearly",
  .long_format = TRUE
)
}
\arguments{
\item{.data}{A data frame containing the data to be converted.}

\item{.country_col}{A character vector specifying the name(s) of the country
identifiers.}

\item{.year_col}{A character vector specifying the name(s) of the time
identifiers.}

\item{.indicator_col}{A character vector specifying the name(s) of the indicator
identifiers.}

\item{.frequency}{A character vector specifying the frequency of the time series.
Default is "yearly" for yearly data. The other values are "month" and "quarter".}

\item{.long_format}{A logical value indicating whether to return the data in long format.
Default is TRUE.}
}
\value{
A tsibble object with the specified country and time identifiers.
}
\description{
This function converts a data frame to a tsibble object.
}
