% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize}
\alias{normalize}
\title{Normalize and detrend the series for each country}
\usage{
normalize(
  .data,
  .cols,
  .country_col = c("Country.Code", "Country.Name"),
  .year_col = "Year",
  .detrend = TRUE,
  .long_format = TRUE
)
}
\arguments{
\item{.data}{A data frame containing the data to be normalized and detrended.}

\item{.cols}{A character vector of column names to be normalized and detrended,
can be integers refering to indexes of the columns (e.g., 1:10), or the names
of the columns (e.g., c("AG.CON.FERT.PT.ZS", "DC.DAC.AUTL.CD")).}

\item{.country_col}{A character vector specifying the name(s) of the country
identifiers}

\item{.year_col}{A character vector specifying the name of the year column.}

\item{.detrend}{A logical value indicating whether to detrend the data. The
\code{supsmu} function is used for smoothing. Default is TRUE.}

\item{.long_format}{A logical value indicating whether to return the data in long format.
Default is TRUE.}
}
\value{
A data frame with the normalized and detrended data.
}
\description{
This function normalizes and detrends the series for each country.
}
