% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{plot.maly_detect}
\alias{plot.maly_detect}
\title{Plot method for maly_detect class}
\usage{
\method{plot}{maly_detect}(x, country = NULL, indicator = NULL, ...)
}
\arguments{
\item{x}{An object of class maly_detect.}

\item{country}{A character vector specifying the country to plot.
Default is NULL, which means first country in the data will be used.}

\item{indicator}{A character vector specifying the indicator to plot.
Default is NULL, which means first indicator in the data will be used.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A ggplot object showing the outliers detected in the specified country and indicator.
}
\description{
This function provides a plot of the outliers detected in the maly_detect class object.
}
