% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{plot.maly_norm}
\alias{plot.maly_norm}
\title{Plot method for maly_norm class}
\usage{
\method{plot}{maly_norm}(x, country = NULL, indicator = NULL, x.lab = NULL, y.lab = NULL, ...)
}
\arguments{
\item{x}{A maly_norm object.}

\item{country}{A character vector specifying the country to plot.
Default is NULL, which means first country in the data will be used.}

\item{indicator}{A character vector specifying the indicator to plot.
Default is NULL, which means first indicator in the data will be used.}

\item{x.lab}{A string specifying the label for the x-axis.
Default is NULL, which means the time column will be used.}

\item{y.lab}{A string specifying the label for the y-axis of the original value (top graph).
Default is NULL, which means the value column will be used.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A ggplot object with the original series and the Zscore of the maly_norm object.
}
\description{
This function plots the original series and the Zscore of the maly_norm object,
for selected country and indicator.
}
